Rename `f_raw_records` to `s_raw_records`
