+++ saltax/plugins/__init__.py
@@ -1,5 +1,5 @@
-from . import f_raw_records
-from .f_raw_records import *
+from . import s_raw_records
+from .s_raw_records import *
 
 from . import records
 from .records import *
